#2 Consolidate conditional with an inline if statement
Tried to consolidate very similar code - the only difference is that one conditional block removes a playlist from selection and the other adds. Everything else was repeated. Replaced a file read with a "with" block as well. Tried to make the code more logical by correcting variable name semantics with "selected_playlist" and "make_active". The ternary expression is hideous though, I would like to eliminate it with some OOP, but I am not sure how as of yet.