#2 Begin transfer from argparse to click (broken atm)
Separated the simplest of options that should be commands. As I cannot think straight right now I refactored this only partly - this version does not work as the original nad is broken as fuck. Will fix it sometime tomorrow - today I simply want to push my code in to start an acitivty streak on Github.